mcp_tools_collection/runcmd_mcp_standalone/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runcmd-mcp"
version = "0.1.4"
description = "A Model Context Protocol (MCP) service for executing system commands asynchronously"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "runcmd-mcp contributors", email = "maintainer@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastmcp>=0.1.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
runcmd-mcp = "runcmd_mcp.__main__:main"

[project.urls]
Homepage = "https://github.com/username/runcmd-mcp"
Repository = "https://github.com/username/runcmd-mcp"
Documentation = "https://github.com/username/runcmd-mcp/blob/main/README.md"

[tool.setuptools.packages.find]
where = ["src"]
include = ["runcmd_mcp*"]

[project.entry-points."mcp.servers"]
runcmd-mcp = "runcmd_mcp.__main__:main"